requirements.txt
# Core model inference
faster-whisper>=1.0.0
llama-cpp-python>=0.3.0
kokoro>=0.9.0

# Audio processing
soundfile>=0.12.0
numpy>=1.24.0

# Pipeline server
websockets>=12.0
PyYAML>=6.0

# Test client (optional — only needed for local testing)
sounddevice>=0.4.6

# Model download (optional — only needed for setup_models.sh)
huggingface_hub>=0.20.0
